Add files via upload
Code designed to run the Tesla Cabin Heater.  It monitors the Fault line for each IGBT, and controls the Reset and IGBT (PWM) lines.  It is designed for an Arduino Mega.  The unit outputs to the Serial port and can take simple commands to start and stop heating.  '?' on the serial gets the software to output the commands.  It is also set up to receive CAN commands to start/stop heating.  The code is not fully tested as I failed to get the hardware to respond - however I'm posting it here as hopefully a good starting point for others.