Documentation of the Tesla Heater Control board
This set of slides is my documentation for the Control Board on the Tesla Cabin Heater.  The initial pins were a starting point (see the YouTube videos in the ReadMe).  I was unable to repeat the method demonstrated in the videos.  I'm not an electrical engineer, so my interpretation of the circuitry on the board could be incorrect.  Please validate before using.